requirements.txt
blinker==1.9.0
Brotli==1.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
ConfigArgParse==1.7
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
locust==2.32.6
MarkupSafe==3.0.2
msgpack==1.1.0
psutil==6.1.1
python-dotenv==1.0.1
pyzmq==26.2.0
requests==2.32.3
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
